refactor(tee): move attestation generation logic from the attestation module
Fixes #706
