+++ attestation/Cargo.toml
@@ -3,10 +3,6 @@ name = "attestation"
 version = "0.1.0"
 edition = "2024"
 
-[features]
-default = []
-external-services-tests = []
-
 [dependencies]
 anyhow = { workspace = true }
 backon = { workspace = true }
