+++ attestation/src/lib.rs
@@ -9,4 +9,3 @@ pub mod measurements;
 pub mod quote;
 pub mod report_data;
 pub mod tcbinfo;
-pub mod tee_authority;
