+++ libs/chain-signatures/contract/Cargo.toml
@@ -65,6 +65,7 @@ k256 = { version = "0.13.4", features = [
 dcap-qvl = { git = "https://github.com/mattlockyer/dcap-qvl", rev = "3caedd8fd0f365841abaf1430ff2922fcb09d9db" }
 
 mpc-primitives = { path = "../../../crates/primitives" }
+attestation = { path = "../../../attestation" }
 near-gas = { version = "0.2.5", features = ["serde", "borsh", "schemars"] }
 near-account-id = "1.1.1"
 thiserror = "1"
