+++ node/src/tee/remote_attestation.rs
@@ -1,187 +1,210 @@
+use crate::indexer::types::{ChainSendTransactionRequest, ProposeJoinArgs};
 use anyhow::{bail, Context};
+use attestation::{
+    attestation::{Attestation, DstackAttestation},
+    collateral::Collateral,
+    measurements::ExpectedMeasurements,
+    quote::Quote,
+    report_data::ReportData,
+    tcbinfo::TcbInfo,
+};
 use backon::{BackoffBuilder, ExponentialBuilder};
-use dstack_sdk::dstack_client::{DstackClient, TcbInfo};
+use core::{future::Future, time::Duration};
+use derive_more::Constructor;
+use dstack_sdk::dstack_client::DstackClient;
 use http::status::StatusCode;
 use mpc_contract::tee::tee_participant::TeeParticipantInfo;
 use near_crypto::PublicKey;
-use reqwest::multipart::Form;
-use serde::{Deserialize, Serialize};
-use sha3::{Digest, Sha3_384};
-use std::{future::Future, time::Duration};
+use reqwest::{multipart::Form, Url};
+use serde::Deserialize;
 use tokio::sync::mpsc;
 use tracing::error;
 
-use crate::indexer::types::{ChainSendTransactionRequest, ProposeJoinArgs};
+#[cfg(test)]
+use attestation::attestation::LocalAttestation;
 
-/// Endpoint to contact dstack service.
-/// Set to [`None`] which defaults to `/var/run/dstack.sock`
-const ENDPOINT: Option<&str> = None;
-/// URL for usbmission of tdx quote. Returns collateral to be used for verification.
-const PHALA_TDX_QUOTE_UPLOAD_URL: &str = "https://proof.t16z.com/api/upload";
-/// Expected HTTP [`StatusCode`] for a successful submission.
-const PHALA_SUCCESS_STATUS_CODE: StatusCode = StatusCode::OK;
-/// The maximum duration to wait for retrying request to Phala's endpoint, [`PHALA_TDX_QUOTE_UPLOAD_URL`].
+/// The maximum duration to wait for retrying request to Phala's endpoint.
 const MAX_BACKOFF_DURATION: Duration = Duration::from_secs(60);
 
-/// Number of bytes for the report data.
-/// report_data: [u8; 64] = [version(2 bytes (big endian)) || sha384(TLS pub key || account public key ) || zero padding]
-const REPORT_DATA_SIZE: usize = 64;
+/// Default URL for submission of TDX quote. Returns collateral to be used for verification.
+const DEFAULT_PHALA_TDX_QUOTE_UPLOAD_URL: &str = "https://proof.t16z.com/api/upload";
 
-const BINARY_VERSION_OFFSET: usize = 0;
-const BINARY_VERSION_SIZE: usize = 2;
+/// Default path for dstack Unix socket endpoint.
+const DEFAULT_DSTACK_ENDPOINT: &str = "/var/run/dstack.sock";
 
-const PUBLIC_KEYS_OFFSET: usize = 2;
-const PUBLIC_KEYS_SIZE: usize = 48;
-
-pub const BINARY_VERSION: BinaryVersion = BinaryVersion(1);
-
-// Compile-time assertions
-const _: () = {
-    assert!(
-        BINARY_VERSION_SIZE + PUBLIC_KEYS_SIZE <= REPORT_DATA_SIZE,
-        "Version and public key must not exceed report data size."
-    );
-    assert!(
-        BINARY_VERSION_OFFSET + BINARY_VERSION_SIZE == PUBLIC_KEYS_OFFSET,
-        "Public key offset must be after binary version."
-    );
-};
-
-#[derive(Serialize, Deserialize)]
-pub struct TeeAttestation {
-    tcb_info: TcbInfo,
-    tdx_quote: String,
-    collateral: serde_json::Value,
+#[cfg(test)]
+#[derive(Constructor)]
+pub struct LocalTeeAuthorityConfig {
+    verification_result: bool,
 }
 
-#[derive(Deserialize)]
-struct UploadResponse {
-    quote_collateral: serde_json::Value,
-    #[serde(rename = "checksum")]
-    _checksum: String,
+#[derive(Constructor)]
+pub struct DstackTeeAuthorityConfig {
+    /// Endpoint to contact dstack service. Defaults to [`DEFAULT_DSTACK_ENDPOINT`]
+    dstack_endpoint: String,
+    /// URL for submission of TDX quote. Returns collateral to be used for verification.
+    quote_upload_url: Url,
 }
 
-pub struct BinaryVersion(u16);
-
-async fn get_with_backoff<Operation, OperationFuture, Value, Error>(
-    operation: Operation,
-    description: &str,
-) -> Value
-where
-    Error: std::fmt::Debug,
-    Operation: Fn() -> OperationFuture,
-    OperationFuture: Future<Output = Result<Value, Error>>,
-{
-    let mut backoff = ExponentialBuilder::default()
-        .with_max_delay(MAX_BACKOFF_DURATION)
-        .without_max_times()
-        .with_jitter()
-        .build();
-
-    // Loop until we have a response.
-    loop {
-        match operation().await {
-            Err(err) => {
-                let duration = backoff.next().unwrap_or(MAX_BACKOFF_DURATION);
-                error!(?err, "{description} failed. retrying in: {:?}", duration);
-                continue;
-            }
-            Ok(response) => break response,
+impl Default for DstackTeeAuthorityConfig {
+    fn default() -> Self {
+        Self {
+            dstack_endpoint: String::from(DEFAULT_DSTACK_ENDPOINT),
+            quote_upload_url: DEFAULT_PHALA_TDX_QUOTE_UPLOAD_URL
+                .parse()
+                .expect("Default URL should be valid"),
         }
     }
 }
 
-/// Generates a [`TeeAttestation`] for this node, which can be used to send to the contract to prove that
-/// the node is running in a `TEE` context.
-///
-/// Returns an [`anyhow::Error`] if a non-transient error occurs, that prevents the node
-/// from generating the attestation.
-pub async fn create_remote_attestation_info(
-    tls_public_key: &PublicKey,
-    account_public_key: &PublicKey,
-) -> TeeAttestation {
-    let client = DstackClient::new(ENDPOINT);
+/// TeeAuthority is an abstraction over different TEE attestation generator implementations. It
+/// generates [`Attestation`] instances - either mocked or real ones.
+pub enum TeeAuthority {
+    Dstack(DstackTeeAuthorityConfig),
+    #[cfg(test)]
+    Local(LocalTeeAuthorityConfig),
+}
 
-    let client_info_response = get_with_backoff(|| client.info(), "dstack client info").await;
-    let tcb_info = client_info_response.tcb_info;
+impl TeeAuthority {
+    /// Generates a [`Attestation`] for this node, which can be used to send to the contract to prove that the node is running in a `TEE` context.
+    ///
+    /// Returns an [`anyhow::Error`] if a non-transient error occurs, that prevents the node from generating the attestation.
+    pub async fn generate_attestation(
+        &self,
+        report_data: ReportData,
+    ) -> anyhow::Result<Attestation> {
+        match self {
+            TeeAuthority::Dstack(config) => {
+                self.generate_dstack_attestation(config, report_data).await
+            }
+            #[cfg(test)]
+            TeeAuthority::Local(config) => Ok(Attestation::Local(LocalAttestation::new(
+                config.verification_result,
+            ))),
+        }
+    }
 
-    let report_data: [u8; REPORT_DATA_SIZE] = {
-        let mut report_data = [0u8; REPORT_DATA_SIZE];
+    async fn generate_dstack_attestation(
+        &self,
+        config: &DstackTeeAuthorityConfig,
+        report_data: ReportData,
+    ) -> anyhow::Result<Attestation> {
+        let client = DstackClient::new(Some(config.dstack_endpoint.as_str()));
 
-        // Copy binary version
-        let byte_representation = BINARY_VERSION.0.to_be_bytes();
-        report_data[BINARY_VERSION_OFFSET..][..BINARY_VERSION_SIZE]
-            .copy_from_slice(&byte_representation);
+        let client_info_response =
+            get_with_backoff(|| client.info(), "dstack client info", None).await?;
+        let tcb_info = TcbInfo::from(client_info_response.tcb_info);
 
-        // Copy hash
-        let mut hasher = Sha3_384::new();
-        hasher.update(tls_public_key.key_data());
-        hasher.update(account_public_key.key_data());
-        let public_keys_hash: [u8; PUBLIC_KEYS_SIZE] = hasher.finalize().into();
-        report_data[PUBLIC_KEYS_OFFSET..][..PUBLIC_KEYS_SIZE].copy_from_slice(&public_keys_hash);
+        let quote = get_with_backoff(
+            || client.get_quote(report_data.to_bytes().into()),
+            "dstack client tdx quote",
+            None,
+        )
+        .await?
+        .quote;
 
-        report_data
-    };
+        let collateral = self
+            .upload_quote_for_collateral(&config.quote_upload_url, &quote)
+            .await?;
+        let quote: Quote = quote.parse()?;
 
-    let tdx_quote: String = get_with_backoff(
-        || client.get_quote(report_data.into()),
-        "dstack client tdx quote",
-    )
-    .await
-    .quote;
+        Ok(Attestation::Dstack(DstackAttestation::new(
+            quote,
+            collateral,
+            tcb_info,
+            ExpectedMeasurements::default(),
+        )))
+    }
 
-    let quote_upload_response = {
-        let reqwest_client = reqwest::Client::new();
-        let tdx_quote = tdx_quote.clone();
+    /// Uploads TDX quote to Phala endpoint and retrieves collateral.
+    async fn upload_quote_for_collateral(
+        &self,
+        quote_upload_url: &Url,
+        tdx_quote: &str,
+    ) -> anyhow::Result<Collateral> {
+        let reqwest_client = reqwest::Client::builder()
+            .timeout(core::time::Duration::from_secs(10))
+            .build()
+            .context("Failed to build HTTP client")?;
+        let tdx_quote = String::from(tdx_quote);
 
         let upload_tdx_quote = async || {
             let form = Form::new().text("hex", tdx_quote.clone());
 
             let response = reqwest_client
-                .post(PHALA_TDX_QUOTE_UPLOAD_URL)
+                .post(quote_upload_url.clone())
                 .multipart(form)
                 .send()
                 .await?;
 
             let status = response.status();
 
-            if status != PHALA_SUCCESS_STATUS_CODE {
-                bail!("Got unexpected HTTP status code: response from phala http_endpoint: {:?}, expected: {:?}", status, PHALA_SUCCESS_STATUS_CODE);
+            if status != StatusCode::OK {
+                bail!(
+                    "Got unexpected HTTP status code: response from phala endpoint: {:?}, expected: {:?}",
+                    status,
+                    StatusCode::OK
+                );
             }
 
-            response
-                .json::<UploadResponse>()
-                .await
-                .context("Failed to deserialize response from Phala.")
+            Ok(response.json::<UploadResponse>().await?)
         };
 
-        get_with_backoff(upload_tdx_quote, "upload tdx quote").await
-    };
+        let upload_response =
+            get_with_backoff(upload_tdx_quote, "upload tdx quote", Some(1)).await?;
 
-    let collateral = quote_upload_response.quote_collateral;
-    TeeAttestation {
-        tdx_quote,
-        tcb_info,
-        collateral,
+        Ok(upload_response.quote_collateral)
     }
 }
 
-impl TryFrom<TeeAttestation> for TeeParticipantInfo {
-    type Error = anyhow::Error;
+#[derive(Deserialize)]
+struct UploadResponse {
+    quote_collateral: Collateral,
+    #[serde(rename = "checksum")]
+    _checksum: String,
+}
 
-    fn try_from(value: TeeAttestation) -> Result<Self, Self::Error> {
-        let tee_quote = hex::decode(value.tdx_quote)
-            .context("Failed to decode tee quote. Expected it to be in hex format.")?;
-        let quote_collateral = serde_json::to_string(&value.collateral)
-            .context("Failed to serialize quote collateral back to JSON string.")?;
-        let raw_tcb_info =
-            serde_json::to_string(&value.tcb_info).context("Failed to serialize tcb info")?;
+async fn get_with_backoff<Operation, OperationFuture, Value, Error>(
+    operation: Operation,
+    description: &str,
+    max_retries: Option<usize>,
+) -> Result<Value, Error>
+where
+    Error: core::fmt::Debug,
+    Operation: Fn() -> OperationFuture,
+    OperationFuture: Future<Output = Result<Value, Error>>,
+{
+    let mut backoff = {
+        let builder = ExponentialBuilder::default()
+            .with_max_delay(MAX_BACKOFF_DURATION)
+            .with_jitter();
 
-        Ok(Self {
-            tee_quote,
-            quote_collateral,
-            raw_tcb_info,
-        })
+        if let Some(max_retries) = max_retries {
+            builder.with_max_times(max_retries).build()
+        } else {
+            builder.without_max_times().build()
+        }
+    };
+
+    // Loop until we have a response or exceed max retries
+    loop {
+        match operation().await {
+            Err(err) => {
+                if let Some(duration) = backoff.next() {
+                    error!(?err, "{description} failed. retrying in: {:?}", duration);
+                    continue;
+                } else {
+                    match max_retries {
+                        Some(retries) => {
+                            error!(?err, "{description} failed after {} retries", retries)
+                        }
+                        None => continue,
+                    }
+                    return Err(err);
+                }
+            }
+            Ok(response) => break Ok(response),
+        }
     }
 }
 
@@ -201,4 +224,109 @@ pub async fn submit_remote_attestation(
         ))
         .await
         .context("Failed to send remote attestation transaction. Channel is closed.")
+}
+
+#[cfg(test)]
+mod tests {
+    use attestation::report_data::ReportDataV1;
+
+    use super::*;
+    use rstest::rstest;
+
+    #[test]
+    fn test_upload_response_deserialization() {
+        // This is a fixture captured from the actual Phala endpoint response
+        let json_response = r#"{
+            "quote_collateral": {
+                "tcb_info_issuer_chain": "-----BEGIN CERTIFICATE-----\nMIICjTCCAjKgAwIBAgIUfjiC1ftVKUpASY5FhAPpFJG99FUwCgYIKoZIzj0EAwIw\naDEaMBgGA1UEAwwRSW50ZWwgU0dYIFJvb3QgQ0ExGjAYBgNVBAoMEUludGVsIENv\ncnBvcmF0aW9uMRQwEgYDVQQHDAtTYW50YSBDbGFyYTELMAkGA1UECAwCQ0ExCzAJ\nBgNVBAYTAlVTMB4XDTI1MDUwNjA5MjUwMFoXDTMyMDUwNjA5MjUwMFowbDEeMBwG\nA1UEAwwVSW50ZWwgU0dYIFRDQiBTaWduaW5nMRowGAYDVQQKDBFJbnRlbCBDb3Jw\nb3JhdGlvbjEUMBIGA1UEBwwLU2FudGEgQ2xhcmExCzAJBgNVBAgMAkNBMQswCQYD\nVQQGEwJVUzBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABENFG8xzydWRfK92bmGv\nP+mAh91PEyV7Jh6FGJd5ndE9aBH7R3E4A7ubrlh/zN3C4xvpoouGlirMba+W2lju\nypajgbUwgbIwHwYDVR0jBBgwFoAUImUM1lqdNInzg7SVUr9QGzknBqwwUgYDVR0f\nBEswSTBHoEWgQ4ZBaHR0cHM6Ly9jZXJ0aWZpY2F0ZXMudHJ1c3RlZHNlcnZpY2Vz\nLmludGVsLmNvbS9JbnRlbFNHWFJvb3RDQS5kZXIwHQYDVR0OBBYEFH44gtX7VSlK\nQEmORYQD6RSRvfRVMA4GA1UdDwEB/wQEAwIGwDAMBgNVHRMBAf8EAjAAMAoGCCqG\nSM49BAMCA0kAMEYCIQDdmmRuAo3qCO8TC1IoJMITAoOEw4dlgEBHzSz1TuMSTAIh\nAKVTqOkt59+co0O3m3hC+v5Fb00FjYWcgeu3EijOULo5\n-----END CERTIFICATE-----\n-----BEGIN CERTIFICATE-----\nMIICjzCCAjSgAwIBAgIUImUM1lqdNInzg7SVUr9QGzknBqwwCgYIKoZIzj0EAwIw\naDEaMBgGA1UEAwwRSW50ZWwgU0dYIFJvb3QgQ0ExGjAYBgNVBAoMEUludGVsIENv\ncnBvcmF0aW9uMRQwEgYDVQQHDAtTYW50YSBDbGFyYTELMAkGA1UECAwCQ0ExCzAJ\nBgNVBAYTAlVTMB4XDTE4MDUyMTEwNDUxMFoXDTQ5MTIzMTIzNTk1OVowaDEaMBgG\nA1UEAwwRSW50ZWwgU0dYIFJvb3QgQ0ExGjAYBgNVBAoMEUludGVsIENvcnBvcmF0\naW9uMRQwEgYDVQQHDAtTYW50YSBDbGFyYTELMAkGA1UECAwCQ0ExCzAJBgNVBAYT\nAlVTMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEC6nEwMDIYZOj/iPWsCzaEKi7\n1OiOSLRFhWGjbnBVJfVnkY4u3IjkDYYL0MxO4mqsyYjlBalTVYxFP2sJBK5zlKOB\nuzCBuDAfBgNVHSMEGDAWgBQiZQzWWp00ifODtJVSv1AbOScGrDBSBgNVHR8ESzBJ\nMEegRaBDhkFodHRwczovL2NlcnRpZmljYXRlcy50cnVzdGVkc2VydmljZXMuaW50\nZWwuY29tL0ludGVsU0dYUm9vdENBLmRlcjAdBgNVHQ4EFgQUImUM1lqdNInzg7SV\nUr9QGzknBqwwDgYDVR0PAQH/BAQDAgEGMBIGA1UdEwEB/wQIMAYBAf8CAQEwCgYI\nKoZIzj0EAwIDSQAwRgIhAOW/5QkR+S9CiSDcNoowLuPRLsWGf/Yi7GSX94BgwTwg\nAiEA4J0lrHoMs+Xo5o/sX6O9QWxHRAvZUGOdRQ7cvqRXaqI=\n-----END CERTIFICATE-----\n",
+                "tcb_info": "{\"id\":\"TDX\",\"version\":3,\"issueDate\":\"2025-07-18T07:00:14Z\",\"nextUpdate\":\"2025-08-17T07:00:14Z\",\"fmspc\":\"20a06f000000\",\"pceId\":\"0000\",\"tcbType\":0,\"tcbEvaluationDataNumber\":17,\"tdxModule\":{\"mrsigner\":\"000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\",\"attributes\":\"0000000000000000\",\"attributesMask\":\"FFFFFFFFFFFFFFFF\"},\"tdxModuleIdentities\":[{\"id\":\"TDX_03\",\"mrsigner\":\"000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\",\"attributes\":\"0000000000000000\",\"attributesMask\":\"FFFFFFFFFFFFFFFF\",\"tcbLevels\":[{\"tcb\":{\"isvsvn\":3},\"tcbDate\":\"2024-03-13T00:00:00Z\",\"tcbStatus\":\"UpToDate\"}]},{\"id\":\"TDX_01\",\"mrsigner\":\"000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\",\"attributes\":\"0000000000000000\",\"attributesMask\":\"FFFFFFFFFFFFFFFF\",\"tcbLevels\":[{\"tcb\":{\"isvsvn\":4},\"tcbDate\":\"2024-03-13T00:00:00Z\",\"tcbStatus\":\"UpToDate\"},{\"tcb\":{\"isvsvn\":2},\"tcbDate\":\"2023-08-09T00:00:00Z\",\"tcbStatus\":\"OutOfDate\"}]}],\"tcbLevels\":[{\"tcb\":{\"sgxtcbcomponents\":[{\"svn\":2,\"category\":\"BIOS\",\"type\":\"Early Microcode Update\"},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"SGX Late Microcode Update\"},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"TXT SINIT\"},{\"svn\":2,\"category\":\"BIOS\"},{\"svn\":2,\"category\":\"BIOS\"},{\"svn\":255,\"category\":\"BIOS\"},{\"svn\":0},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"SEAMLDR ACM\"},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0}],\"pcesvn\":13,\"tdxtcbcomponents\":[{\"svn\":5,\"category\":\"OS/VMM\",\"type\":\"TDX Module\"},{\"svn\":0,\"category\":\"OS/VMM\",\"type\":\"TDX Module\"},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"TDX Late Microcode Update\"},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0}]},\"tcbDate\":\"2024-03-13T00:00:00Z\",\"tcbStatus\":\"UpToDate\"},{\"tcb\":{\"sgxtcbcomponents\":[{\"svn\":2,\"category\":\"BIOS\",\"type\":\"Early Microcode Update\"},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"SGX Late Microcode Update\"},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"TXT SINIT\"},{\"svn\":2,\"category\":\"BIOS\"},{\"svn\":2,\"category\":\"BIOS\"},{\"svn\":255,\"category\":\"BIOS\"},{\"svn\":0},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"SEAMLDR ACM\"},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0}],\"pcesvn\":5,\"tdxtcbcomponents\":[{\"svn\":5,\"category\":\"OS/VMM\",\"type\":\"TDX Module\"},{\"svn\":0,\"category\":\"OS/VMM\",\"type\":\"TDX Module\"},{\"svn\":2,\"category\":\"OS/VMM\",\"type\":\"TDX Late Microcode Update\"},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0},{\"svn\":0}]},\"tcbDate\":\"2018-01-04T00:00:00Z\",\"tcbStatus\":\"OutOfDate\"}]}",
+                "tcb_info_signature": "6f1a8292d9b0457f7207126043e7110fbbf2455da844626cd9d070ff25f2f20302a1d58320f032c6ae162b9e2d0ba7c549ab90ae52fa66bc757c9ce362a79bc1",
+                "qe_identity_issuer_chain": "-----BEGIN CERTIFICATE-----\nMIICjTCCAjKgAwIBAgIUfjiC1ftVKUpASY5FhAPpFJG99FUwCgYIKoZIzj0EAwIw\naDEaMBgGA1UEAwwRSW50ZWwgU0dYIFJvb3QgQ0ExGjAYBgNVBAoMEUludGVsIENv\ncnBvcmF0aW9uMRQwEgYDVQQHDAtTYW50YSBDbGFyYTELMAkGA1UECAwCQ0ExCzAJ\nBgNVBAYTAlVTMB4XDTI1MDUwNjA5MjUwMFoXDTMyMDUwNjA5MjUwMFowbDEeMBwG\nA1UEAwwVSW50ZWwgU0dYIFRDQiBTaWduaW5nMRowGAYDVQQKDBFJbnRlbCBDb3Jw\nb3JhdGlvbjEUMBIGA1UEBwwLU2FudGEgQ2xhcmExCzAJBgNVBAgMAkNBMQswCQYD\nVQQGEwJVUzBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABENFG8xzydWRfK92bmGv\nP+mAh91PEyV7Jh6FGJd5ndE9aBH7R3E4A7ubrlh/zN3C4xvpoouGlirMba+W2lju\nypajgbUwgbIwHwYDVR0jBBgwFoAUImUM1lqdNInzg7SVUr9QGzknBqwwUgYDVR0f\nBEswSTBHoEWgQ4ZBaHR0cHM6Ly9jZXJ0aWZpY2F0ZXMudHJ1c3RlZHNlcnZpY2Vz\nLmludGVsLmNvbS9JbnRlbFNHWFJvb3RDQS5kZXIwHQYDVR0OBBYEFH44gtX7VSlK\nQEmORYQD6RSRvfRVMA4GA1UdDwEB/wQEAwIGwDAMBgNVHRMBAf8EAjAAMAoGCCqG\nSM49BAMCA0kAMEYCIQDdmmRuAo3qCO8TC1IoJMITAoOEw4dlgEBHzSz1TuMSTAIh\nAKVTqOkt59+co0O3m3hC+v5Fb00FjYWcgeu3EijOULo5\n-----END CERTIFICATE-----\n-----BEGIN CERTIFICATE-----\nMIICjzCCAjSgAwIBAgIUImUM1lqdNInzg7SVUr9QGzknBqwwCgYIKoZIzj0EAwIw\naDEaMBgGA1UEAwwRSW50ZWwgU0dYIFJvb3QgQ0ExGjAYBgNVBAoMEUludGVsIENv\ncnBvcmF0aW9uMRQwEgYDVQQHDAtTYW50YSBDbGFyYTELMAkGA1UECAwCQ0ExCzAJ\nBgNVBAYTAlVTMB4XDTE4MDUyMTEwNDUxMFoXDTQ5MTIzMTIzNTk1OVowaDEaMBgG\nA1UEAwwRSW50ZWwgU0dYIFJvb3QgQ0ExGjAYBgNVBAoMEUludGVsIENvcnBvcmF0\naW9uMRQwEgYDVQQHDAtTYW50YSBDbGFyYTELMAkGA1UECAwCQ0ExCzAJBgNVBAYT\nAlVTMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEC6nEwMDIYZOj/iPWsCzaEKi7\n1OiOSLRFhWGjbnBVJfVnkY4u3IjkDYYL0MxO4mqsyYjlBalTVYxFP2sJBK5zlKOB\nuzCBuDAfBgNVHSMEGDAWgBQiZQzWWp00ifODtJVSv1AbOScGrDBSBgNVHR8ESzBJ\nMEegRaBDhkFodHRwczovL2NlcnRpZmljYXRlcy50cnVzdGVkc2VydmljZXMuaW50\nZWwuY29tL0ludGVsU0dYUm9vdENBLmRlcjAdBgNVHQ4EFgQUImUM1lqdNInzg7SV\nUr9QGzknBqwwDgYDVR0PAQH/BAQDAgEGMBIGA1UdEwEB/wQIMAYBAf8CAQEwCgYI\nKoZIzj0EAwIDSQAwRgIhAOW/5QkR+S9CiSDcNoowLuPRLsWGf/Yi7GSX94BgwTwg\nAiEA4J0lrHoMs+Xo5o/sX6O9QWxHRAvZUGOdRQ7cvqRXaqI=\n-----END CERTIFICATE-----\n",
+                "qe_identity": "{\"id\":\"TD_QE\",\"version\":2,\"issueDate\":\"2025-07-18T06:50:31Z\",\"nextUpdate\":\"2025-08-17T06:50:31Z\",\"tcbEvaluationDataNumber\":17,\"miscselect\":\"00000000\",\"miscselectMask\":\"FFFFFFFF\",\"attributes\":\"11000000000000000000000000000000\",\"attributesMask\":\"FBFFFFFFFFFFFFFF0000000000000000\",\"mrsigner\":\"DC9E2A7C6F948F17474E34A7FC43ED030F7C1563F1BABDDF6340C82E0E54A8C5\",\"isvprodid\":2,\"tcbLevels\":[{\"tcb\":{\"isvsvn\":4},\"tcbDate\":\"2024-03-13T00:00:00Z\",\"tcbStatus\":\"UpToDate\"}]}",
+                "qe_identity_signature": "cb2e6fcdb5a521295cb27817285e01b58482674e8c55b61702c3a8e8eb708c50dd29fbe164e1cfda9d66b475cd8e19c52c3c1f85fa613a305cec4ba237835a4c"
+            },
+            "checksum": "878cb2e2150f4f5274b5a0b57dd0124555b625dc432c298537bc62eeb9f7fe8a"
+        }"#;
+
+        let upload_response: UploadResponse = serde_json::from_str(json_response).unwrap();
+        let collateral = upload_response.quote_collateral;
+
+        assert!(!collateral.tcb_info_issuer_chain.is_empty());
+        assert!(collateral
+            .tcb_info_issuer_chain
+            .contains("-----BEGIN CERTIFICATE-----"));
+
+        assert!(!collateral.tcb_info.is_empty());
+        assert!(collateral.tcb_info.contains("TDX"));
+
+        assert!(!collateral.tcb_info_signature.is_empty());
+        assert_eq!(collateral.tcb_info_signature.len(), 64);
+
+        assert!(!collateral.qe_identity_issuer_chain.is_empty());
+        assert!(collateral
+            .qe_identity_issuer_chain
+            .contains("-----BEGIN CERTIFICATE-----"));
+
+        assert!(!collateral.qe_identity.is_empty());
+        assert!(collateral.qe_identity.contains("TD_QE"));
+
+        assert!(!collateral.qe_identity_signature.is_empty());
+        assert_eq!(collateral.qe_identity_signature.len(), 64);
+    }
+
+    #[rstest]
+    #[tokio::test]
+    async fn test_generate_and_verify_attestation_local(
+        #[values(true, false)] quote_verification_result: bool,
+    ) {
+        let tls_key = "ed25519:DcA2MzgpJbrUATQLLceocVckhhAqrkingax4oJ9kZ847"
+            .parse()
+            .unwrap();
+        let account_key = "ed25519:H9k5eiU4xXyb8F7cUDjZYNuH1zGAx5BBNrYwLPNhq6Zx"
+            .parse()
+            .unwrap();
+        let report_data = ReportData::V1(ReportDataV1::new(tls_key, account_key));
+
+        let authority =
+            TeeAuthority::Local(LocalTeeAuthorityConfig::new(quote_verification_result));
+        let attestation = authority
+            .generate_attestation(report_data.clone())
+            .await
+            .unwrap();
+        let timestamp_s = 0u64;
+        assert_eq!(
+            attestation.verify(report_data, timestamp_s, &[]),
+            quote_verification_result
+        );
+    }
+
+    #[tokio::test]
+    #[cfg(feature = "external-services-tests")]
+    async fn test_upload_quote_for_collateral_with_phala_endpoint() {
+        extern crate std;
+
+        const VALID_QUOTE_HEX: &str = "040002008100000000000000939a7233f79c4ca9940a0db3957f0607ac666ed993e70e31ff5f5a8a2c743b220000000007010300000000000000000000000000c51e5cb16c461fe29b60394984755325ecd05a9a7a8fb3a116f1c3cf0aca4b0eb9edefb9b404deeaee4b7d454372d17a000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000010000000000702000000000000c68518a0ebb42136c12b2275164f8c72f25fa9a34392228687ed6e9caeb9c0f1dbd895e9cf475121c029dc47e70e91fd00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000085e0855a6384fa1c8a6ab36d0dcbfaa11a5753e5a070c08218ae5fe872fcb86967fd2449c29e22e59dc9fec998cb65474a7db64a609c77e85f603c23e9a9fd03bfd9e6b52ce527f774a598e66d58386026cea79b2aea13b81a0b70cfacdec0ca8a4fe048fea22663152ef128853caa5c033cbe66baf32ba1ff7f6b1afc1624c279f50a4cbc522a735ca6f69551e61ef2561c1b02351cd6f7c803dd36bc95ba25463aa025ce7761156260c9131a5d7c03aeccc10e12160ec3205bb2876a203a7fb81447910d62fd92897d68b1f51d54fb75dfe2aeba3a97a879cba59a771fc522d88046cc26b407d723f726fae17c3e5a50529d0b6c2b991d027f06a9b430d43ecc1000003bdd12b68ee3cfc93a1758479840b6f8734c2439106d8f0faa50ac919d86ea101c002c41d262670ad84afb8f9ee35c7abbb72dcc01bbc3e3a3773672d665005ee6bcb0c5f4b03f0563c797747f7ddd25d92d4f120bee4a829daca986bbc03c155b3d158f6a386bca7ee49ceb3ec31494b792e0cf22fc4e561ddc57156da1b77a0600461000000303070704ff00020000000000000000000000000000000000000000000000000000000000000000000000000000000015000000000000000700000000000000e5a3a7b5d830c2953b98534c6c59a3a34fdc34e933f7f5898f0a85cf08846bca0000000000000000000000000000000000000000000000000000000000000000dc9e2a7c6f948f17474e34a7fc43ed030f7c1563f1babddf6340c82e0e54a8c5000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000020006000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000005d2eb8ae211693884eadaea0be0392c5532c7ff55429e4696c84954444d62ed600000000000000000000000000000000000000000000000000000000000000004f1cd2dde7dd5d4a9a495815f3ac76c56a77a9e06a5279a8c8550b54cf2d7287a630c3b9aefb94b1b6e8491eba4b43baa811c8f44167eb7d9ca933678ea64f5b2000000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f05005e0e00002d2d2d2d2d424547494e2043455254494649434154452d2d2d2d2d0a4d49494538544343424a656741774942416749554439426b736e734170713045567861464a59785a56794f6774664d77436759494b6f5a497a6a3045417749770a634445694d434147413155454177775a535735305a577767553064594946424453794251624746305a6d397962534244515445614d42674741315545436777520a535735305a577767513239796347397959585270623234784644415342674e564241634d43314e68626e526849454e7359584a684d51737743515944565151490a44414a445154454c4d416b474131554542684d4356564d774868634e4d6a55774d6a41334d5463774f4441325768634e4d7a49774d6a41334d5463774f4441320a576a42774d534977494159445651514444426c4a626e526c624342545231676755454e4c49454e6c636e52705a6d6c6a5958526c4d526f77474159445651514b0a4442464a626e526c6243424462334a7762334a6864476c76626a45554d424947413155454277774c553246756447456751327868636d4578437a414a42674e560a4241674d416b4e424d517377435159445651514745774a56557a425a4d424d4742797147534d34394167454743437147534d34394177454841304941424853770a3977506a72554532734f4a644c5653415434686565414a572b31796c6473615556696b5a4c485832506235777374326a79697539414f5865576a7a6a6d585a4c0a4343742b457858716f53394e45476c6b52724b6a67674d4e4d4949444354416642674e5648534d4547444157674253566231334e765276683655424a796454300a4d383442567776655644427242674e56485238455a4442694d47436758714263686c706f64485277637a6f764c32467761533530636e567a6447566b633256790a646d6c6a5a584d75615735305a577775593239744c334e6e6543396a5a584a3061575a7059324630615739754c3359304c33426a61324e796244396a595431770a624746305a6d397962535a6c626d4e765a476c755a7a316b5a584977485159445652304f42425945464d6a464e59626f7464634b636859487258467966774b460a774e534d4d41344741315564447745422f775145417749477744414d42674e5648524d4241663845416a41414d4949434f67594a4b6f5a496876684e415130420a424949434b7a4343416963774867594b4b6f5a496876684e41513042415151514134346b35686a336951797044574873756f5a474144434341575147436971470a534962345451454e41514977676746554d42414743797147534962345451454e41514942416745434d42414743797147534962345451454e41514943416745430a4d42414743797147534962345451454e41514944416745434d42414743797147534962345451454e41514945416745434d42414743797147534962345451454e0a41514946416745434d42454743797147534962345451454e41514947416749412f7a415142677371686b69472b453042445145434277494241444151426773710a686b69472b4530424451454343414942416a415142677371686b69472b45304244514543435149424144415142677371686b69472b45304244514543436749420a4144415142677371686b69472b45304244514543437749424144415142677371686b69472b45304244514543444149424144415142677371686b69472b4530420a44514543445149424144415142677371686b69472b45304244514543446749424144415142677371686b69472b453042445145434477494241444151426773710a686b69472b45304244514543454149424144415142677371686b69472b45304244514543455149424454416642677371686b69472b45304244514543456751510a4167494341674c2f4141494141414141414141414144415142676f71686b69472b45304244514544424149414144415542676f71686b69472b453042445145450a424159676f473841414141774477594b4b6f5a496876684e4151304242516f424154416542676f71686b69472b453042445145474242414b496f456755387a650a486d2b49596f7a686c337a314d45514743697147534962345451454e415163774e6a415142677371686b69472b45304244514548415145422f7a4151426773710a686b69472b45304244514548416745422f7a415142677371686b69472b45304244514548417745422f7a414b42676771686b6a4f5051514441674e49414442460a4169417362735a44796d2f72455a30476c454c62442f6e64755061536a485341746e5871567453313047486255774968414d585666784b334b666f4b675131660a4578397478765331314362363662323467424344523963477942562b0a2d2d2d2d2d454e442043455254494649434154452d2d2d2d2d0a2d2d2d2d2d424547494e2043455254494649434154452d2d2d2d2d0a4d4949436c6a4343416a32674177494241674956414a567658633239472b487051456e4a3150517a7a674658433935554d416f4743437147534d343942414d430a4d476778476a415942674e5642414d4d45556c756447567349464e48574342536232393049454e424d526f77474159445651514b4442464a626e526c624342440a62334a7762334a6864476c76626a45554d424947413155454277774c553246756447456751327868636d4578437a414a42674e564241674d416b4e424d5173770a435159445651514745774a56557a4165467730784f4441314d6a45784d4455774d5442614677307a4d7a41314d6a45784d4455774d5442614d484178496a41670a42674e5642414d4d47556c756447567349464e4857434251513073675547786864475a76636d306751304578476a415942674e5642416f4d45556c75644756730a49454e76636e4276636d4630615739754d5251774567594456515148444174545957353059534244624746795954454c4d416b474131554543417743513045780a437a414a42674e5642415954416c56544d466b77457759484b6f5a497a6a3043415159494b6f5a497a6a304441516344516741454e53422f377432316c58534f0a3243757a7078773734654a423732457944476757357258437478327456544c7136684b6b367a2b5569525a436e71523770734f766771466553786c6d546c4a6c0a65546d693257597a33714f42757a43427544416642674e5648534d4547444157674251695a517a575770303069664f44744a5653763141624f536347724442530a42674e5648523845537a424a4d45656752614244686b466f64485277637a6f764c324e6c636e52705a6d6c6a5958526c63793530636e567a6447566b633256790a646d6c6a5a584d75615735305a577775593239744c306c756447567355306459556d397664454e424c6d526c636a416442674e5648513445466751556c5739640a7a62306234656c4153636e553944504f4156634c336c517744675944565230504151482f42415144416745474d42494741315564457745422f7751494d4159420a4166384341514177436759494b6f5a497a6a30454177494452774177524149675873566b6930772b6936565947573355462f32327561586530594a446a3155650a6e412b546a44316169356343494359623153416d4435786b66545670766f34556f79695359787244574c6d5552344349394e4b7966504e2b0a2d2d2d2d2d454e442043455254494649434154452d2d2d2d2d0a2d2d2d2d2d424547494e2043455254494649434154452d2d2d2d2d0a4d4949436a7a4343416a53674177494241674955496d554d316c71644e496e7a6737535655723951477a6b6e42717777436759494b6f5a497a6a3045417749770a614445614d4267474131554541777752535735305a5777675530645949464a766233516751304578476a415942674e5642416f4d45556c756447567349454e760a636e4276636d4630615739754d5251774567594456515148444174545957353059534244624746795954454c4d416b47413155454341774351304578437a414a0a42674e5642415954416c56544d423458445445344d4455794d5445774e4455784d466f58445451354d54497a4d54497a4e546b314f566f77614445614d4267470a4131554541777752535735305a5777675530645949464a766233516751304578476a415942674e5642416f4d45556c756447567349454e76636e4276636d46300a615739754d5251774567594456515148444174545957353059534244624746795954454c4d416b47413155454341774351304578437a414a42674e56424159540a416c56544d466b77457759484b6f5a497a6a3043415159494b6f5a497a6a3044415163445167414543366e45774d4449595a4f6a2f69505773437a61454b69370a314f694f534c52466857476a626e42564a66566e6b59347533496a6b4459594c304d784f346d717379596a6c42616c54565978465032734a424b357a6c4b4f420a757a43427544416642674e5648534d4547444157674251695a517a575770303069664f44744a5653763141624f5363477244425342674e5648523845537a424a0a4d45656752614244686b466f64485277637a6f764c324e6c636e52705a6d6c6a5958526c63793530636e567a6447566b63325679646d6c6a5a584d75615735300a5a577775593239744c306c756447567355306459556d397664454e424c6d526c636a416442674e564851344546675155496d554d316c71644e496e7a673753560a55723951477a6b6e4271777744675944565230504151482f42415144416745474d42494741315564457745422f7751494d4159424166384341514577436759490a4b6f5a497a6a3045417749445351417752674968414f572f35516b522b533943695344634e6f6f774c7550524c735747662f59693747535839344267775477670a41694541344a306c72486f4d732b586f356f2f7358364f39515778485241765a55474f6452513763767152586171493d0a2d2d2d2d2d454e442043455254494649434154452d2d2d2d2d0a0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000";
+
+        let tee_authority = TeeAuthority::Dstack(DstackTeeAuthorityConfig::default());
+        let config = DstackTeeAuthorityConfig::default();
+
+        let result = tokio::time::timeout(
+            std::time::Duration::from_secs(10),
+            tee_authority.upload_quote_for_collateral(&config.quote_upload_url, VALID_QUOTE_HEX),
+        )
+        .await;
+
+        match result {
+            Ok(Ok(collateral)) => {
+                assert!(!collateral.tcb_info_issuer_chain.is_empty());
+                assert!(!collateral.tcb_info.is_empty());
+                assert!(!collateral.tcb_info_signature.is_empty());
+                assert!(!collateral.qe_identity_issuer_chain.is_empty());
+                assert!(!collateral.qe_identity.is_empty());
+                assert!(!collateral.qe_identity_signature.is_empty());
+            }
+            Ok(Err(e)) => panic!("Test failed: {e:?}"),
+            Err(e) => panic!("Test timed out: {e:?}"),
+        }
+    }
 }
